Refactor auth provider, UserMenu and PodLoginPage
